Some XML tests.
Currently, all this does is check whether we can prepend a commented
line to ZFREFLE1.MIZ, compute the abstract, and see that it is the
same as the abstract for the real ZFREFLE1.MIZ.

I've sketched some further tests. I'm not especially fond of the name
"xml.t".  Later on, we change it.  The code from these tests should
eventually go into our pre-commit hook (or some other utility file).