A "smart" makefile for generating MML XML and PREL.
It's "smart" in the sense that it keeps track of when the prel's for
articles are generated.  One can then make some changes to the MML
and, assuming that the article's environment is unchanged, this
makefile can be used to call miz2prel and mizf on only those articles
that depend on the changed article.

Next up: deal with cases where the the article environment changes.  I
suppose we should just store the dependencies as a perl hash, from
which this thing can be generated.