2 line killer; Not sure about the scary .PHONY targets: the only advantage of makefile over the setup script is that it can be run not just on the beginning, but economically for various stages (e.g. commiting new library version to "stable"). But the .PHONIES will kill any work-saving as we have seen before.
MAKEJOBS should be read from the environment.
The REPO-compiled is missing - not good, this is our cache of a cleanly
compiling state.
Not sure about the frontend permissions; in the backend, we need a
full user for the various ops, and we should keep the master repo at
mizarw now - that's in our paper.
The "html" target looks like a repetition of the Makefile.repo targets.