Some new test data for the itemizer.
This file contains the first few items from the RAT_1 article.  It's a
challenge because it involves proof-like structure at the
top-level (i.e., not contined withing a proof).  One has to recognize
that the first item,

  0 in omega;

stands alone, and then to realize that the second item

  then reconsider 09 = 0 as Element of REAL+ by ARYTM_2:2;

is justified by the first item (that's what the initial "then" is
doing).  At the same time, this is a top-leve reconsider item, which
I'm not sure how to handle yet.  In the third item, there are uses of
the newly introduced term 09.