Put the LibXML use statement at the beginning.
Is there any advantage to deferring the use statement until we need
it?  Maybe not.  I didn't put the use statement at the beginning of
the file because I wanted to load the module only when we had
determined that the options were coherent; otherwise, loading the XML
module is a waste, since it's not going to be used.

Probably I'm just being ignorant about perl and modules here.